Add !strict parsing
